LayoutTests/imported/w3c/web-platform-tests/html/infrastructure/urls/resolving-urls/query-encoding/windows-1252_include=svg-expected.txt


PASS SVG <a>
TIMEOUT SVG <feImage> Test timed out
FAIL SVG <image> assert_equals: expected "%E5" but got "%C3%A5"
PASS SVG <use>
